#CONGTACDOAN-6 feat: show tcList popup
